refactor(mcp): share workspace identifier routing helper
