+++ src/basic_memory/mcp/tools/edit_note.py
@@ -7,19 +7,22 @@ from loguru import logger
 from fastmcp import Context
 from pydantic import AliasChoices, Field
 
-from basic_memory.config import BasicMemoryConfig, ConfigManager
+from basic_memory.config import ConfigManager
 from basic_memory.mcp.project_context import (
     _cloud_workspace_discovery_available,
     detect_project_from_memory_url_prefix,
     get_project_client,
     add_project_metadata,
     resolve_project_and_path,
-    resolve_workspace_qualified_identifier,
 )
 from basic_memory.mcp.server import mcp
 from basic_memory.schemas.base import Entity
 from basic_memory.schemas.response import EntityResponse
-from basic_memory.utils import normalize_project_reference, validate_project_path
+from basic_memory.services.link_resolver import (
+    detect_project_from_workspace_identifier_prefix,
+    is_workspace_qualified_plain_identifier,
+)
+from basic_memory.utils import validate_project_path
 
 
 def _parse_identifier_to_title_and_directory(identifier: str) -> tuple[str, str]:
@@ -47,37 +50,6 @@ def _parse_identifier_to_title_and_directory(identifier: str) -> tuple[str, str]
         title = cleaned
 
     return title, directory
-
-
-def _is_workspace_qualified_plain_identifier(identifier: str) -> bool:
-    """Return True for plain ``<workspace>/<project>/<path>`` identifiers."""
-    stripped = identifier.strip()
-    if stripped.startswith("memory://"):
-        return False
-
-    normalized = normalize_project_reference(stripped).strip("/")
-    return len(normalized.split("/", 2)) == 3
-
-
-async def detect_project_from_workspace_identifier_prefix(
-    identifier: str,
-    config: BasicMemoryConfig,
-    context: Context | None = None,
-) -> Optional[str]:
-    """Resolve a project route from a plain workspace-qualified identifier."""
-    if not _is_workspace_qualified_plain_identifier(identifier):
-        return None
-
-    if not _cloud_workspace_discovery_available(config):
-        return None
-
-    workspace_resolution = await resolve_workspace_qualified_identifier(
-        identifier,
-        context=context,
-    )
-    if workspace_resolution is None:
-        return None
-    return workspace_resolution.project_identifier
 
 
 def _format_error_response(
@@ -338,7 +310,7 @@ async def edit_note(
             )
         elif _cloud_workspace_discovery_available(
             config
-        ) and _is_workspace_qualified_plain_identifier(identifier):
+        ) and is_workspace_qualified_plain_identifier(identifier):
             detected = await detect_project_from_workspace_identifier_prefix(
                 identifier,
                 config,
